requirements.txt
google-genai
faiss-cpu
numpy
pydantic
python-dotenv